Add output file option to cli
Adds an option for specifying an output file when using markdown-it from the command line.

I'm on the VScode team and in our docs we currently recomend that users install marked for converting markdown to html: https://code.visualstudio.com/Docs/languages/markdown#_compiling-markdown-into-html . I'd like to use markdown-it in this example instead, but correctly setting up a task for this is not as easy as it is with marked. Having a commandline output file option addresses this.